file_handler.py documented
